Commit 29/05: After a month the project is changed a lot.
Now multiple users are managed.

A token authorization scheme has been implemented. Users, at the time of
the login, will be provided with a JWT token. Since then, all the
requests must be attached with a authorization header field in the form
of "Bearer "+token. The token is used by the system for both
authentication and authorization. The TokenManager (now simply
implemented by a java class) keeps record of tokens, corresponding users
and their allowed files (owned or shared with). No expiration scheme is
implemented now, even though it will be a TODO in following releases.
The token management will be transferred to a key-value in memory
datastore (e.g. Redis).

The file, now, contains information about its owner and other users it
is shared with. However, the sharing features are not implemented yet
(TODO).

The middleware model has been separated from the api model (both for
performance and security reason). Finally, a TestServices class has been
provided; it can be used as test for the services but also as use case
samples.